requirements.txt
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
pdfminer.six==20220524
pycparser==2.21